test: extract shared sweep fixtures into tests/_promote_fixtures.py
Post-b8 sweep SonarCloud quality gate flagged new_duplicated_lines_
density at 10.95% (allowed 3%) on commit 9c82fb9. The 59 duplicated
lines were three small fixtures (`_make_simple_handler`,
`_fake_find_title_match`, `_run_promote_simple`) copy-pasted across
the b6/b7/b8 sweep test files.

Extract them into a single shared module ``tests/_promote_fixtures.py``
(underscore prefix → not collected by pytest's
``python_files = ["test_*.py"]``). Update b6, b7, b8 to import the
helpers under the same underscore-prefixed local aliases the existing
call sites use. No behaviour change.

b6's sweep-specific helpers (``_archive_stub``, ``_fake_title_match_hit``,
``_run_promote_synthesize``) stay local — those drive the synthesize-
path and aren't shared with b7/b8.

## Tests

```
2448 passed, 54 skipped (full suite, ~28s)
```

mypy clean across 52 source files. black + flake8 clean.
